Deploys new personal website
Replaces the old theme with a new custom layout for a personal website.

The new layout includes sections for about, skills, experience,
projects, and contact information, with a modern design and
responsive layout.

Removes git submodule update command from the amplify.yml since it is no
longer needed and updates the base URL in config.yaml.